# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/addtocart": {
            "get": {
                "description": "It will just add the product in the user cart",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart Apis"
                ],
                "summary": "to add in cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/admin/addproduct": {
            "post": {
                "description": "you can add products here by filling product details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Apis"
                ],
                "summary": "To add a product in the store",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/cartcheckout": {
            "get": {
                "description": "It will just buy the product from the user cart",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart Apis"
                ],
                "summary": "to buy from cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/instantbuy": {
            "get": {
                "description": "It will just buy the product instantly",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart Apis"
                ],
                "summary": "to buy instant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "userid",
                        "name": "userid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pid",
                        "name": "pid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/removeitem": {
            "get": {
                "description": "It will just remove the product from the user cart",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart Apis"
                ],
                "summary": "to remove in cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Please provide email and password to signin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Apis"
                ],
                "summary": "User SignIn",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/users/productview": {
            "get": {
                "description": "It will give you all the products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Apis"
                ],
                "summary": "To show all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/users/search": {
            "get": {
                "description": "It will give you the product for product id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Apis"
                ],
                "summary": "To get product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "description": "Please provide firstname, lastname, phone, email and password to signup",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Apis"
                ],
                "summary": "User SignUp",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Address": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "string"
                },
                "city_name": {
                    "type": "string"
                },
                "house_name": {
                    "type": "string"
                },
                "pin_code": {
                    "type": "string"
                },
                "street_name": {
                    "type": "string"
                }
            }
        },
        "models.Order": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "integer"
                },
                "order_ID": {
                    "type": "string"
                },
                "order_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductUser"
                    }
                },
                "ordered_on": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.Payment"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
        "models.Payment": {
            "type": "object",
            "properties": {
                "cod": {
                    "type": "boolean"
                },
                "digital": {
                    "type": "boolean"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "product_ID": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "models.ProductUser": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "product_ID": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password",
                "phone"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Address"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                },
                "refresh_Token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updtaed_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "usercart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductUser"
                    }
                }
            }
        },
        "responses.ApplicationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}